requirements.txt
#python -m venv venv

# 3. Activate it
#venv\Scripts\activate

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Hugging Face - Medical Models
transformers>=4.36.0
torch>=2.2.0
sentence-transformers>=2.3.0

# Document Processing (NEW)
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0
python-multipart>=0.0.6

# Utilities
numpy>=1.26.0